Controllers Refactor
- Create User removed
- Create User logic moved directly inside registerUser (renamed from registerUserBuyer)
- Removed Testing controllers

+++ back/DietiDeals24_25/src/main/java/com/dietideals/dietideals24_25/controllers/AuctionController.java
@@ -62,7 +62,7 @@ public class AuctionController {
             AuctionEntity savedAuctionEntity = auctionService.save(auctionEntity);
             AuctionDto responseAuction = auctionMapper.mapTo(savedAuctionEntity);
 
-            if (itemExists) {
+            if (itemExists && item != null) {
                 item.setAuction(savedAuctionEntity);
                 item = itemService.save(item);
                 responseAuction.setItem(itemMapper.mapTo(item));
